FIX: Move flask object creation code from __init__.py to app.py
