+++ test/test_bucketlist.py
@@ -3,7 +3,7 @@
 """
 import unittest
 import json
-from app import create_app, db
+from app.app import create_app, db
 
 class BucketlistTestCase(unittest.TestCase):
     """Test cases for bucketlists
